transformed header view to css modules :/
