+++ package.json
@@ -13,8 +13,8 @@
     "install-docusaurus": "yalc add @docusaurus/core & yarn install"
   },
   "dependencies": {
-    "@agile-ts/core": "0.0.10",
-    "@agile-ts/react": "0.0.10",
+    "@agile-ts/core": "0.0.11",
+    "@agile-ts/react": "0.0.11",
     "@docusaurus/core": "^2.0.0-alpha.70",
     "@docusaurus/module-type-aliases": "^2.0.0-alpha.70",
     "@docusaurus/preset-classic": "^2.0.0-alpha.70",
