+++ src/components/buttons/PrimaryButton/styles.module.css
@@ -1,0 +1,21 @@
+.ButtonContainer {
+  color: var(--ifm-color-on-primary);
+  background-color: var(--ifm-color-primary);
+  border-radius: 3px;
+  border: none;
+
+  font-size: 18px;
+  font-weight: bold;
+
+  cursor: pointer;
+
+  padding: 15px 30px;
+  align-items: center;
+  justify-content: center;
+
+  transition: 0.3s ease all;
+}
+
+.ButtonContainer:hover {
+  background-color: var(--ifm-color-primary-lighter);
+}
