+++ src/components/other/HeaderTyper/AutoTyper/events/Event.ts
