+++ src/components/other/HeaderTyper/AutoTyper/events/LoopEvent.ts
