+++ src/components/other/HeaderTyper/AutoTyper/events/RemoveEvent.ts
