+++ src/components/other/HeaderTyper/AutoTyper/events/SleepEvent.ts
