+++ src/components/other/HeaderTyper/AutoTyper/events/TypeEvent.ts
