+++ src/components/other/HeaderTyper/AutoTyper/index.ts
