+++ src/pages/LandingPage/components/HeaderView/styles.module.css
@@ -1,0 +1,104 @@
+.Container {
+    display: flex;
+    height: 500px;
+    width: 100%;
+
+    margin: 0 auto;
+    max-width: var(--ifm-container-width);
+    padding: 4.5rem 1rem;
+    margin-bottom: 200px;
+}
+
+
+.AstronautImage {
+    position: absolute;
+    right: 0;
+    width: 45%;
+    max-width: 800px;
+    opacity: 1;
+}
+
+.HeaderTitle {
+    color: var(--ifm-navbar-link-color);
+    font-size: 4rem;
+    font-weight: bold;
+}
+
+.SeparatorContainer {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+}
+
+.SeparatorLeft {
+    background-color: var(--ifm-navbar-link-active-color);
+    width: 20px;
+    height: 2px;
+}
+
+.SeparatorText {
+    color: var(--ifm-navbar-link-active-color);
+    font-size: 1.25rem;
+    margin-left: 15px;
+    margin-right: 15px;
+    margin-bottom: 7px;
+}
+
+.SeparatorRight {
+    background-color: var(--ifm-navbar-link-active-color);
+    width: 220px;
+    height: 2px;
+}
+
+.DescriptionText {
+    color: var(--ifm-color-primary-lightest);
+    font-size: 1.5rem;
+    max-width: 400px;
+}
+
+.ButtonContainer {
+    display: flex;
+    flex-direction: row;
+}
+
+.GithubButton {
+    width: 200px;
+}
+
+.GetStartedButton {
+    width: 200px;
+    margin-right: 20px;
+}
+
+@media (max-width: 996px) {
+    .AstronautImage {
+        opacity: 0;
+    }
+}
+
+@media (max-width: 768px) {
+    .Container {
+        margin-top: 50px;
+    }
+
+    .HeaderTitle {
+        font-size: 2rem;
+    }
+
+    .SeparatorRight {
+        width: 150px;
+    }
+
+    .DescriptionText {
+        font-size: 1.3rem;
+    }
+
+    .ButtonContainer {
+        flex-direction: column;
+    }
+
+    .GetStartedButton {
+        margin-right: 0;
+        margin-bottom: 20px;
+    }
+}
